+++ packages/web-design/src/components/fleet-pi/pi/settings-dialog.tsx
@@ -386,6 +386,7 @@ function SettingsDialogSession({
   }
 
   const handleDiscardChanges = () => {
+    revertResourceDraft()
     resetDraft()
     resetCommittedModelBaseline()
     setDiscardDialogOpen(false)